# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserSortKey": {
				"type": "string",
				"enum": [
					"name",
					"email",
					"createdAt",
					"role"
				]
			},
			"ObjectId": {
				"type": "string",
				"description": "A class representation of the BSON ObjectId type."
			},
			"SocialLink": {
				"properties": {
					"platform": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"platform",
					"url"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Pick_User.Exclude_keyofUser.password__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"admin",
							"superadmin"
						]
					},
					"phone": {
						"type": "string"
					},
					"orders": {
						"items": {
							"$ref": "#/components/schemas/ObjectId"
						},
						"type": "array"
					},
					"addresses": {
						"items": {
							"$ref": "#/components/schemas/ObjectId"
						},
						"type": "array"
					},
					"authorize": {
						"type": "boolean"
					},
					"googleId": {
						"type": "string"
					},
					"facebookId": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"profile": {
						"properties": {
							"socialLinks": {
								"items": {
									"$ref": "#/components/schemas/SocialLink"
								},
								"type": "array"
							},
							"description": {
								"type": "string"
							},
							"title": {
								"type": "string"
							},
							"photo": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"email",
					"authorize"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.password_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.password__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateUserRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"admin"
						]
					},
					"authorize": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"phone",
					"authorize"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UpdateUserRequest": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"admin",
							"superadmin"
						]
					},
					"authorize": {
						"type": "boolean"
					},
					"profile": {
						"properties": {
							"socialLinks": {
								"items": {
									"$ref": "#/components/schemas/SocialLink"
								},
								"type": "array"
							},
							"description": {
								"type": "string"
							},
							"title": {
								"type": "string"
							},
							"photo": {
								"type": "string"
							}
						},
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"ProductContent": {
				"properties": {
					"description": {
						"type": "string"
					},
					"feature": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"care": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"description",
					"feature",
					"care"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ProductVariant": {
				"properties": {
					"color": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"size": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"color",
					"size"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ProductMedia": {
				"properties": {
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"video": {
						"type": "string"
					}
				},
				"required": [
					"images",
					"video"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CreateProductRequest": {
				"properties": {
					"sku": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"product_type": {
						"type": "string"
					},
					"partner": {
						"type": "string",
						"nullable": true
					},
					"set": {
						"type": "string",
						"nullable": true
					},
					"content": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ProductContent"
							}
						],
						"nullable": true
					},
					"variant": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ProductVariant"
							}
						],
						"nullable": true
					},
					"media": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ProductMedia"
							}
						],
						"nullable": true
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"nullable": true
					},
					"sale_price": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						],
						"nullable": true
					},
					"sale_period": {
						"type": "string",
						"nullable": true
					},
					"parent": {
						"items": {
							"type": "string",
							"nullable": true
						},
						"type": "array"
					},
					"children": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"nullable": true
					},
					"recommended": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"sku",
					"name",
					"price",
					"product_type",
					"partner",
					"set",
					"content",
					"variant",
					"media",
					"tags",
					"sale_price",
					"sale_period",
					"parent",
					"children",
					"recommended"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Product": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"sku": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"product_type": {
						"type": "string"
					},
					"partner": {
						"type": "string",
						"nullable": true
					},
					"set": {
						"type": "string",
						"nullable": true
					},
					"content": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ProductContent"
							}
						],
						"nullable": true
					},
					"variant": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ProductVariant"
							}
						],
						"nullable": true
					},
					"media": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ProductMedia"
							}
						],
						"nullable": true
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"nullable": true
					},
					"sale_price": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						],
						"nullable": true
					},
					"sale_period": {
						"type": "string",
						"nullable": true
					},
					"parent": {
						"items": {
							"type": "string",
							"nullable": true
						},
						"type": "array",
						"nullable": true
					},
					"children": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"nullable": true
					},
					"recommended": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"sku",
					"name",
					"price",
					"product_type",
					"partner",
					"set",
					"content",
					"variant",
					"media",
					"tags",
					"sale_price",
					"sale_period",
					"parent",
					"children",
					"recommended"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ProductSortKey": {
				"type": "string",
				"enum": [
					"sku",
					"name",
					"price",
					"product_type",
					"partner",
					"set",
					"sale_price",
					"sale_period",
					"createdAt",
					"updatedAt"
				]
			},
			"SortOrder": {
				"type": "string",
				"enum": [
					"asc",
					"desc"
				]
			},
			"UpdateProductRequest": {
				"properties": {
					"sku": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"product_type": {
						"type": "string"
					},
					"partner": {
						"type": "string"
					},
					"set": {
						"type": "string"
					},
					"content": {
						"$ref": "#/components/schemas/ProductContent"
					},
					"variant": {
						"$ref": "#/components/schemas/ProductVariant"
					},
					"media": {
						"$ref": "#/components/schemas/ProductMedia"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sale_price": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					},
					"sale_period": {
						"type": "string"
					},
					"parent": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"children": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"recommended": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "app",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/v1/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "filters",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UserSortKey"
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"default": "asc",
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {},
										"token": {
											"type": "string"
										},
										"user": {
											"$ref": "#/components/schemas/Omit_User.password_"
										}
									},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserRequest"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Omit_User.password_"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserRequest"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/{id}/upload-photo": {
			"post": {
				"operationId": "UploadPhoto",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
...
		"/api/v1/products": {
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllProducts",
				"responses": {
					"200": {
						"description": "Products fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"skip": {
											"type": "number",
											"format": "double"
										},
										"limit": {
											"type": "number",
											"format": "double"
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"products": {
											"items": {
												"$ref": "#/components/schemas/Product"
											},
											"type": "array"
										}
									},
									"required": [
										"skip",
										"limit",
										"total",
										"products"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ProductSortKey"
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortOrder"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "product_type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tags",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/products/{id}": {
			"get": {
				"operationId": "GetProductById",
				"responses": {
					"200": {
						"description": "Product fetched",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Product updated successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProductRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Product deleted successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}